+++ include/llvm/Transforms/Tapir/TapirUtils.h
@@ -1,0 +1,58 @@
+//===- TapirUtils.h - Tapir Helper functions                ----*- C++ -*--===//
+//
+//                     The LLVM Compiler Infrastructure
+//
+// This file is distributed under the University of Illinois Open Source
+// License. See LICENSE.TXT for details.
+//
+//===----------------------------------------------------------------------===//
+//
+// This pass is a simple pass wrapper around the PromoteMemToReg function call
+// exposed by the Utils library.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef TAPIR_UTILS_H_
+#define TAPIR_UTILS_H_
+
+#include "llvm/ADT/SmallPtrSet.h"
+#include "llvm/ADT/SmallVector.h"
+#include "llvm/IR/Dominators.h"
+#include "llvm/IR/Instructions.h"
+
+namespace llvm {
+namespace tapir {
+
+bool verifyDetachedCFG(const DetachInst &Detach, DominatorTree &DT,
+                       bool error = true);
+
+bool populateDetachedCFG(const DetachInst &Detach, DominatorTree &DT,
+                         SmallPtrSetImpl<BasicBlock *> &functionPieces,
+                         SmallVectorImpl<BasicBlock *> &reattachB,
+                         SmallPtrSetImpl<BasicBlock *> &ExitBlocks,
+                         bool replace, bool error = true);
+
+Function *extractDetachBodyToFunction(DetachInst &Detach,
+                                     DominatorTree &DT, AssumptionCache &AC,
+                                     CallInst **call = nullptr);
+
+class TapirTarget {
+
+public:
+  //! For use in loopspawning grainsize calculation
+  virtual Value *GetOrCreateWorker8(Function &F) = 0;
+
+  virtual void createSync(SyncInst &inst, ValueToValueMapTy &DetachCtxToStackFrame) = 0;
+
+  virtual Function *createDetach(DetachInst &Detach,
+                         ValueToValueMapTy &DetachCtxToStackFrame,
+                         DominatorTree &DT, AssumptionCache &AC) = 0;
+  virtual void preProcessFunction(Function &F) = 0;
+  virtual void postProcessFunction(Function &F) = 0;
+  virtual void postProcessHelper(Function &F) = 0;
+};
+
+} // ns: tapir
+} // ns: llvm
+
+#endif
