+++ lib/Transforms/Tapir/CMakeLists.txt
@@ -9,6 +9,7 @@ add_llvm_library(LLVMTapirOpts
   LoopSpawning.cpp
   Outline.cpp
   Tapir.cpp
+  TapirUtils.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm/Transforms
