+++ lib/Transforms/Tapir/LowerToCilk.cpp
@@ -17,32 +17,29 @@
 #include "llvm/IR/Verifier.h"
 #include "llvm/Transforms/Utils/Cloning.h"
 #include "llvm/Transforms/Tapir.h"
+#include "llvm/Transforms/Tapir/TapirUtils.h"
 
 #define DEBUG_TYPE "tapir2cilk"
 
 using namespace llvm;
-
-static cl::opt<bool> ClInstrumentCilk("instrument-cilk", cl::init(false),
-                                      cl::Hidden,
-                                      cl::desc("Instrument Cilk events"));
-
-cl::opt<bool> fastCilk("fast-cilk", cl::init(false), cl::Hidden,
-                       cl::desc("Attempt faster cilk call implementation"));
+using namespace llvm::tapir;
 
 namespace {
 
-struct LowerTapirToCilk : public ModulePass {
+struct LowerTapirToTarget : public ModulePass {
   static char ID; // Pass identification, replacement for typeid
   bool DisablePostOpts;
   bool Instrument;
-  explicit LowerTapirToCilk(bool DisablePostOpts = false, bool Instrument = false)
+  TapirTarget* tapirTarget;
+  explicit LowerTapirToTarget(bool DisablePostOpts = false, bool Instrument = false)
       : ModulePass(ID), DisablePostOpts(DisablePostOpts),
         Instrument(Instrument) {
-    initializeLowerTapirToCilkPass(*PassRegistry::getPassRegistry());
+    tapirTarget = new CilkABI(Instrument);
+    initializeLowerTapirToTargetPass(*PassRegistry::getPassRegistry());
   }
 
   StringRef getPassName() const override {
-    return "Simple Lowering of Tapir to Cilk ABI";
+    return "Simple Lowering of Tapir to Target ABI";
   }
 
   bool runOnModule(Module &M) override;
@@ -59,42 +56,16 @@ private:
 };
 }  // End of anonymous namespace
 
-char LowerTapirToCilk::ID = 0;
-INITIALIZE_PASS_BEGIN(LowerTapirToCilk, "tapir2cilk",
+char LowerTapirToTarget::ID = 0;
+INITIALIZE_PASS_BEGIN(LowerTapirToTarget, "tapir2cilk",
                       "Simple Lowering of Tapir to Cilk ABI", false, false)
 INITIALIZE_PASS_DEPENDENCY(AssumptionCacheTracker)
 INITIALIZE_PASS_DEPENDENCY(DominatorTreeWrapperPass)
-INITIALIZE_PASS_END(LowerTapirToCilk, "tapir2cilk",
+INITIALIZE_PASS_END(LowerTapirToTarget, "tapir2cilk",
                     "Simple Lowering of Tapir to Cilk ABI", false, false)
 
-// Helper function to inline calls to compiler-generated Cilk Plus runtime
-// functions when possible.  This inlining is necessary to properly implement
-// some Cilk runtime "calls," such as __cilkrts_detach().
-static inline void inlineCilkFunctions(Function &F) {
-  bool inlining = true;
-  while (inlining) {
-    inlining = false;
-    for (inst_iterator I = inst_begin(F), E = inst_end(F); I != E; ++I)
-      if (CallInst *cal = dyn_cast<CallInst>(&*I))
-        if (Function *fn = cal->getCalledFunction())
-          if (fn->getName().startswith("__cilk")) {
-            InlineFunctionInfo IFI;
-            if (InlineFunction(cal, IFI)) {
-              if (fn->getNumUses()==0)
-                fn->eraseFromParent();
-              inlining = true;
-              break;
-            }
-          }
-  }
 
-  if (verifyFunction(F, &errs())) {
-    DEBUG(F.dump());
-    assert(0);
-  }
-}
-
-bool LowerTapirToCilk::unifyReturns(Function &F) {
+bool LowerTapirToTarget::unifyReturns(Function &F) {
   SmallVector<BasicBlock *, 4> ReturningBlocks;
   for (BasicBlock &BB : F)
     if (isa<ReturnInst>(BB.getTerminator()))
@@ -133,17 +104,13 @@ bool LowerTapirToCilk::unifyReturns(Function &F) {
 }
 
 SmallVectorImpl<Function *>
-*LowerTapirToCilk::processFunction(Function &F, DominatorTree &DT,
+*LowerTapirToTarget::processFunction(Function &F, DominatorTree &DT,
                                    AssumptionCache &AC) {
-  if (fastCilk && F.getName()=="main") {
-    IRBuilder<> start(F.getEntryBlock().getFirstNonPHIOrDbg());
-    auto m = start.CreateCall(CILKRTS_FUNC(init, *F.getParent()));
-    m->moveBefore(F.getEntryBlock().getTerminator());
-  }
-
   if (unifyReturns(F))
     DT.recalculate(F);
 
+  tapirTarget->preProcessFunction(F);
+  
   // Lower Tapir instructions in this function.  Collect the set of helper
   // functions generated by this process.
   SmallVector<Function *, 4> *NewHelpers = new SmallVector<Function *, 4>();
@@ -151,13 +118,11 @@ SmallVectorImpl<Function *>
     if (DetachInst* DI = dyn_cast_or_null<DetachInst>(I->getTerminator())) {
       // Lower a detach instruction, and collect the helper function generated
       // in this process for executing the detached task.
-      Function *Helper = cilk::createDetach(*DI, DetachCtxToStackFrame, DT, AC,
-                                            ClInstrumentCilk || Instrument);
+      Function *Helper = tapirTarget->createDetach(*DI, DetachCtxToStackFrame, DT, AC);
       NewHelpers->push_back(Helper);
     } else if (SyncInst* SI = dyn_cast_or_null<SyncInst>(I->getTerminator())) {
       // Lower a sync instruction.
-      cilk::createSync(*SI, DetachCtxToStackFrame,
-                       ClInstrumentCilk || Instrument);
+      tapirTarget->createSync(*SI, DetachCtxToStackFrame);
     }
   }
 
@@ -166,15 +131,15 @@ SmallVectorImpl<Function *>
     assert(0);
   }
 
-  // Inline Cilk runtime calls in the function and generated helper functions.
-  inlineCilkFunctions(F);
-  for (Function *H : *NewHelpers)
-    inlineCilkFunctions(*H);
+  tapirTarget->postProcessFunction(F);
+  for (Function *H : *NewHelpers) {
+    tapirTarget->postProcessHelper(*H);
+  }
 
   return NewHelpers;
 }
 
-bool LowerTapirToCilk::runOnModule(Module &M) {
+bool LowerTapirToTarget::runOnModule(Module &M) {
   if (skipModule(M))
     return false;
 
@@ -210,10 +175,10 @@ bool LowerTapirToCilk::runOnModule(Module &M) {
   return Changed;
 }
 
-// createLowerTapirToCilkPass - Provide an entry point to create this pass.
+// createLowerTapirToTargetPass - Provide an entry point to create this pass.
 //
 namespace llvm {
-ModulePass *createLowerTapirToCilkPass(bool DisablePostOpts, bool Instrument) {
-  return new LowerTapirToCilk(DisablePostOpts, Instrument);
+ModulePass *createLowerTapirToTargetPass(bool DisablePostOpts, bool Instrument) {
+  return new LowerTapirToTarget(DisablePostOpts, Instrument);
 }
 }
